pyproject.toml
[project]
name = "uao-desarrollo-de-proyectos-de-ia"
version = "0.2.0"
description = "Proyecto final del curso Desarrollo de Proyectos de Inteligencia Artificial"
readme = "README.md"
license = "MIT"
license-files = ["LICENSE"]
authors = [
    {name = "Alexander Calambas Ramirez", email = "alexandercalambas23@gmail.com"},
    {name = "Angelo Parra Cortez", email = "angeloparra207@gmail.com"},
    {name = "Oscar Portela Ospina", email = "osport726@gmail.com"},
    {name = "Sebastian Torres Cabrera", email = "sebastian.torres0426@gmail.com"},
]
classifiers = [
    "Topic :: Education",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
requires-python = ">=3.10,<3.13"
dependencies = [
    "torch>=2.0",
    "torchvision>=0.15",
    "transformers>=4.40",
    "Pillow>=10.0",
    "scikit-learn>=1.3",
    "numpy>=1.24",
    "joblib>=1.3",
    "fastapi>=0.110",
    "uvicorn[standard]>=0.24",
    "python-multipart>=0.0.9",
    "jinja2>=3.1",
    "grpcio>=1.60",
    "protobuf>=4.25",
    "mlflow>=2.10",
    "reportlab>=4.4.10",
]

[tool.uv]
torch-backend = "cpu"

[dependency-groups]
dev = [
    "pre-commit>=4.5.1",
    "pytest>=8.0",
    "httpx>=0.27",
    "ruff>=0.15.2",
    "grpcio-tools>=1.60",
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
addopts = "--verbose"
filterwarnings = [
    "ignore::DeprecationWarning",
]

[tool.ruff]
src = ["src"]
fix = false
output-format = "full"

[tool.ruff.lint]
select = ["B", "E", "F", "I", "UP", "W"]

[tool.ruff.lint.isort]
force-single-line = true
order-by-type = false
